requirements.txt
alabaster==1.0.0
asgiref==3.8.1
babel==2.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.5
django-environ==0.12.0
django-filter==24.3
djangorestframework==3.15.2
docutils==0.21.2
drf-yasg==1.21.8
idna==3.10
imagesize==1.4.1
inflection==0.5.1
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
psycopg2-binary==2.9.10
Pygments==2.19.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
